pyproject.toml
[tool.poetry]
name = "uoapi"
version = "0.2.0"
description = "A robust Python client for interacting with the Newcastle Urban Observatory API (v1.1). This package provides a comprehensive interface for accessing and analyzing urban sensor data with built-in caching, configuration management, and data analysis capabilities."
authors = ["carrowmw <c.morris-wiltshire@newcastle.ac.uk>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
requests = "^2.32.3"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
mkdocs = "^1.6.1"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"